feat: - rename ResNet to CNNBase & rename CNNBase to CNNBaseLegacy - optimize import with relative
