+++ zae_engine/models/builds/__init__.py
@@ -1,0 +1,2 @@
+from .cnn import CNNBase
+from .legacy import ResNet1D, Regressor1D, Segmentor, CNNBaseLegacy
