+++ zae_engine/models/foundations/resnet.py
@@ -1,8 +1,8 @@
 from importlib import import_module
-from typing import Type, OrderedDict
+from typing import OrderedDict
 
-from ..builds.resnet import ResNet
-from ..builds.blocks.resblock import BasicBlock, Bottleneck
+from ..builds.cnn import CNNBase
+from ..blocks.resblock import BasicBlock, Bottleneck
 
 res_map = {
     18: {"block": BasicBlock, "layers": [2, 2, 2, 2]},
@@ -48,7 +48,7 @@ def weight_mapper(src_weight: [OrderedDict | dict], dst_weight: [OrderedDict | d
 
 
 def resnet18(pretrained=False):
-    model = ResNet(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[18])
+    model = CNNBase(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[18])
     if pretrained:
         src_weight = import_module("torchvision.models").ResNet18_Weights.IMAGENET1K_V1.get_state_dict(True)
         dst_weight = weight_mapper(src_weight, model.state_dict())
@@ -57,7 +57,7 @@ def resnet18(pretrained=False):
 
 
 def resnet34(pretrained=False):
-    model = ResNet(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[34])
+    model = CNNBase(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[34])
     if pretrained:
         src_weight = import_module("torchvision.models").ResNet34_Weights.IMAGENET1K_V1.get_state_dict(True)
         dst_weight = weight_mapper(src_weight, model.state_dict())
@@ -66,7 +66,7 @@ def resnet34(pretrained=False):
 
 
 def resnet50(pretrained=False):
-    model = ResNet(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[50])
+    model = CNNBase(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[50])
     if pretrained:
         src_weight = import_module("torchvision.models").ResNet50_Weights.IMAGENET1K_V1.get_state_dict(True)
         dst_weight = weight_mapper(src_weight, model.state_dict())
@@ -75,7 +75,7 @@ def resnet50(pretrained=False):
 
 
 def resnet101(pretrained=False):
-    model = ResNet(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[101])
+    model = CNNBase(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[101])
     if pretrained:
         src_weight = import_module("torchvision.models").ResNet101_Weights.IMAGENET1K_V1.get_state_dict(True)
         dst_weight = weight_mapper(src_weight, model.state_dict())
@@ -84,7 +84,7 @@ def resnet101(pretrained=False):
 
 
 def resnet152(pretrained=False):
-    model = ResNet(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[152])
+    model = CNNBase(ch_in=3, width=64, n_cls=1000, groups=1, dilation=1, **res_map[152])
     if pretrained:
         src_weight = import_module("torchvision.models").ResNet152_Weights.IMAGENET1K_V1.get_state_dict(True)
         dst_weight = weight_mapper(src_weight, model.state_dict())
